Don't try to dowload/install cypress
foreman_ansible depends on rex and tasks. all three, via their NPM deps,
somehow end up pulling in cypress, which fails spectacularly to install
when three processes try to install it at the same time. luckilly, we
don't need cypress at all to build assets, so we can just force it not
to try installation and save us all the pain and security issues it
brings.